Move PageLinkHandler to a submodule wagtail.core.rich_text.pages
